refactored the code to have controllers
